services/remote/requirements.txt
paramiko
python-dotenv

